+++ test/visual.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { runTestSuites } from './compare-screenshots';
 import { allSuites } from './definitions';
+import { runTestSuites } from './definitions/visual/compare-screenshots';
 
 runTestSuites(allSuites);
