--- test/visual/definitions/alert.ts
@@ -1,37 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { TestSuite } from '../types';
-
-const suite: TestSuite = {
-  description: 'alert',
-  tests: [
-    {
-      description: 'simple',
-      path: 'alert/simple',
-      screenshotType: 'screenshotArea',
-    },
-    {
-      description: 'style custom page',
-      path: 'alert/style-custom-types',
-      screenshotType: 'screenshotArea',
-    },
-    ...[600, 1280].map(width => ({
-      description: `width ${width}px`,
-      tests: [
-        {
-          description: 'permutations',
-          path: 'alert/permutations',
-          screenshotType: 'permutations' as const,
-        },
-        {
-          description: 'custom types',
-          path: 'alert/style-custom-types',
-          screenshotType: 'screenshotArea' as const,
-        },
-      ],
-    })),
-  ],
-};
-
-export default suite;
